refactor(core): split internals by responsibility
